requirements.txt
asgiref==3.10.0
boto3==1.40.69
botocore==1.40.69
Django==5.2.8
django-storages==1.14.6
jmespath==1.0.1
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
python-dotenv==1.2.1
